--- src/factories/factoryInterfaces.ts
@@ -1,8 +1,0 @@
-import { BusData, RampAccessSchedule, Stop, RouteShape } from "@/types/mapTypes";
-
-export interface IStmFactory{
-    createBusData(data:JSON) : BusData[] | null
-    createRampAccessSchedule(data:JSON) : RampAccessSchedule[] | null
-    createRouteShape(data:JSON) : RouteShape | null
-    createStops(data:JSON) : Stop[] | null
-}
